Rename `MessageValidationRenderer` to `*Injector`
