+++ java/src/main/kotlin/io/spine/validation/java/RequiredGenerator.kt
@@ -48,9 +48,9 @@ import io.spine.validation.UnsetValue
 import io.spine.validation.java.ErrorPlaceholder.FIELD_PATH
 import io.spine.validation.java.ErrorPlaceholder.FIELD_TYPE
 import io.spine.validation.java.ErrorPlaceholder.PARENT_TYPE
-import io.spine.validation.java.MessageValidationRenderer.MessageScope.message
-import io.spine.validation.java.MessageValidationRenderer.ValidateScope.parentPath
-import io.spine.validation.java.MessageValidationRenderer.ValidateScope.violations
+import io.spine.validation.java.ValidationCodeInjector.MessageScope.message
+import io.spine.validation.java.ValidationCodeInjector.ValidateScope.parentPath
+import io.spine.validation.java.ValidationCodeInjector.ValidateScope.violations
 
 /**
  * The generator for `(required)` option.
