+++ java/src/main/kotlin/io/spine/validation/java/RuleGenerator.kt
@@ -37,7 +37,7 @@ import io.spine.server.query.Querying
 import io.spine.server.query.select
 import io.spine.validation.CompilationMessage
 import io.spine.validation.Rule
-import io.spine.validation.java.MessageValidationRenderer.ValidateScope.violations
+import io.spine.validation.java.ValidationCodeInjector.ValidateScope.violations
 import com.squareup.javapoet.CodeBlock as PoetCodeBlock
 import com.squareup.javapoet.FieldSpec as PoetField
 import com.squareup.javapoet.MethodSpec as PoetMethod
@@ -47,7 +47,7 @@ import com.squareup.javapoet.MethodSpec as PoetMethod
  *
  * Despite implementing [OptionGenerator], this generator is not responsible
  * for a specific option. It generates code for constraints represented by `Rule`.
- * Having rule generation as [OptionGenerator] allows [MessageValidationRenderer] treat
+ * Having rule generation as [OptionGenerator] allows [JavaValidationRenderer] treat
  * all generators in a unified manner, significantly simplifying the class implementation.
  */
 @Deprecated("Will be removed as we complete our migration to standalone generators.")
